pyproject.toml
[project]
name = "shibboleth-threat-detection"
version = "3.0.0"
description = "A risk based authentication implementation designed for Shibboleth IdP"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "geoip2>=5.2.0",
    "gunicorn>=23.0.0",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "mlflow>=3.9.0",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pandas-stubs>=3.0.0.260204",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy-stubs>=1.17.0.2",
    "sqlalchemy>=2.0.46",
    "torch>=2.10.0",
    "user-agents>=2.2.0",
    "waitress>=3.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["web", "nn_scripts", "db", "helpers", "external_data"]

[project.scripts]
api = "web.app:main"
api-prod = "web.run_api:main"
train = "nn_scripts.train_nn:main"
explain = "nn_scripts.explain_nn:main"
verify = "helpers.verify_config_usage:main"
